requirements.txt
albumentations==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blobfile==2.0.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.11.1
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
-e git+https://github.com/openai/improved-diffusion.git@783b6740edb79fdb7d063250db2c51cc9545dcd1#egg=improved_diffusion
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.1
mat4py==0.5.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mpi4py==3.1.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
prdc==0.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.16.0
Pygments==2.14.0
pynndescent==0.5.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-ignite==0.4.10
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.13.1
torch-fidelity==0.3.0
torchmetrics==0.11.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
umap==0.1.1
umap-learn==0.5.3
urllib3==1.26.14
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
zipp==3.11.0
